+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Fixed
+
+- Fixed duplicate entries in `.factory_sloth_done` file
+
 ## [1.2.1] - 2023-05-17
 
 ### Fixed